Edited full.sh and minimal-server.sh

+++ minimal-server.sh
@@ -22,14 +22,20 @@ echo "\nConfiguring symlinks..."
 ln -fs $BASHRC $HOME/.bashrc
 ln -fs $VIMRC $HOME/.vimrc
 
-# Install Oh My Zsh!
-sh -c "$(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)"
+read -p "Install ZSH? (y/n) " confirmation
 
-# Create symbolic link for Headline Oh My Zsh! Theme
-ln -s $ZSH_THEME/headline.zsh-theme ~/.oh-my-zsh/custom/themes/headline.zsh-theme
+if [[ "$confirmation" == [yY]* ]]; then
+	# Install Oh My Zsh!
+	sh -c "$(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)"
 
-# Add ZSH to /etc/shells
-echo "\nAdding ZSH to allowed shells..."
-command -v zsh | sudo tee -a /etc/shells
+	# Create symbolic link for Headline Oh My Zsh! Theme
+	ln -s $ZSH_THEME/headline.zsh-theme ~/.oh-my-zsh/custom/themes/headline.zsh-theme
 
-echo "Run chsh to change shell to ZSH."
+	# Add ZSH to /etc/shells
+	command -v zsh | sudo tee -a /etc/shells
+
+	echo "Run chsh to change shell to ZSH."
+	echo "Complete."
+else
+	echo "ZSH installation aborted."
+fi
